Move instance type to _types/instance.py
Also, include a docstring for the Instance dataclass
