+++ tamr_client/_types/__init__.py
@@ -16,6 +16,7 @@ from tamr_client._types.attribute import (
     SubAttribute,
 )
 from tamr_client._types.dataset import AnyDataset, Dataset, UnifiedDataset
+from tamr_client._types.instance import Instance
 from tamr_client._types.json import JsonDict
 from tamr_client._types.operation import Operation
 from tamr_client._types.project import MasteringProject, Project
