+++ tamr_client/_types/url.py
@@ -1,7 +1,6 @@
 from dataclasses import dataclass
 
-from tamr_client import instance
-from tamr_client.instance import Instance
+from tamr_client._types.instance import Instance
 
 
 @dataclass(frozen=True)
@@ -11,5 +10,7 @@ class URL:
     base_path: str = "api/versioned/v1"
 
     def __str__(self):
+        from tamr_client import instance
+
         origin = instance.origin(self.instance)
         return f"{origin}/{self.base_path}/{self.path}"
