+++ tamr_client/dataset/unified.py
@@ -4,8 +4,7 @@ See https://docs.tamr.com/reference/dataset-models
 from copy import deepcopy
 
 from tamr_client import operation, response
-from tamr_client._types import JsonDict, Operation, UnifiedDataset, URL
-from tamr_client.instance import Instance
+from tamr_client._types import Instance, JsonDict, Operation, UnifiedDataset, URL
 from tamr_client.project import Project
 from tamr_client.session import Session
