+++ tamr_client/project.py
@@ -1,6 +1,5 @@
 from tamr_client import response
-from tamr_client._types import JsonDict, Project, URL
-from tamr_client.instance import Instance
+from tamr_client._types import Instance, JsonDict, Project, URL
 from tamr_client.mastering import project as mastering_project
 from tamr_client.session import Session
